feat: use jpa audit to handle creation and update dates on user, rental, message

+++ src/main/java/com/openclassrooms/chatop/listener/RentalListener.java
@@ -4,8 +4,11 @@ import com.openclassrooms.chatop.exceptions.StorageFileNotFoundException;
 import com.openclassrooms.chatop.model.Rental;
 import com.openclassrooms.chatop.service.FileService;
 import jakarta.persistence.PostRemove;
+import jakarta.persistence.PreUpdate;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+
+import java.time.LocalDateTime;
 
 public class RentalListener {
 
@@ -19,7 +22,7 @@ public class RentalListener {
 
     @PostRemove
     public void  postRemove(Rental rental) {
-        if(!rental.getPicture().isBlank()) {
+        if(rental.getPicture() != null && !rental.getPicture().isBlank()) {
             logger.info("Removing picture for rental {}", rental.getPicture());
             try {
                 fileService.deleteFileFromUrl(rental.getPicture());
